+++ CHANGELOG.md
@@ -1,9 +1,30 @@
+# [1.3.0-alpha.3](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/compare/v1.3.0-alpha.2...v1.3.0-alpha.3) (2021-12-31)
+
+
+### Bug Fixes
+
+* :bug: list active versions after publishing the new version ([#27](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/issues/27)) ([2ae2370](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/commit/2ae2370115d63caf2a9079de6e1a6f9e9a07a10f))
+
 ## [1.2.3](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/compare/v1.2.2...v1.2.3) (2021-12-31)
 
 
 ### Bug Fixes
 
 * :bug: list active versions after publishing the new version ([#27](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/issues/27)) ([2ae2370](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/commit/2ae2370115d63caf2a9079de6e1a6f9e9a07a10f))
+
+# [1.3.0-alpha.2](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/compare/v1.3.0-alpha.1...v1.3.0-alpha.2) (2021-12-31)
+
+
+### Bug Fixes
+
+* :lock: don't write the auth token in npmrc ([#31](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/issues/31)) ([7669fce](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/commit/7669fce5aee784788ddd480151af63519902fd8b))
+
+# [1.3.0-alpha.1](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/compare/v1.2.2...v1.3.0-alpha.1) (2021-12-31)
+
+
+### Features
+
+* :lock: check that authentication is correctly set up before publishing ([#30](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/issues/30)) ([883fd07](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/commit/883fd0730249ddc87687b0dff3a073d0e96aef91))
 
 ## [1.2.2](https://github.com/ghusse/semantic-release-npm-deprecate-old-versions/compare/v1.2.1...v1.2.2) (2021-12-31)